[#31] Refactor : 댓글 컴포넌트 리팩토링
- 불필요한 코드 삭제 처리
- CommentList를 컨테이너화 하여 불필요한 props drilling 제거
