+++ frontend/src/containers/comment/CommentViewerContainer.js
@@ -9,7 +9,7 @@ import {
 	readRecomments,
 	writeComment,
 } from '../../modules/comment/commentActions';
-import CommentsList from '../../components/comment/CommentsList';
+import CommentsList from '../comment/CommentsList';
 import CommentEditor from '../../components/comment/CommentEditor';
 import PaginationBar from '../../components/pagination/PaginationBar';
 
@@ -19,12 +19,10 @@ const CommentViewerContainer = () => {
 	const dispatch = useDispatch();
 
 	const comments = useSelector((state) => state.comments.comments);
-	const recomments = useSelector((state) => state.comments.recomments);
 	const addedComment = useSelector((state) => state.comments.addedComment);
 	const deletedComment = useSelector((state) => state.comments.deletedComment);
 	const error = useSelector((state) => state.comments.error);
 	const loading = useSelector((state) => state.loading['comment/READ_COMMENTS']);
-	const currentUserId = useSelector((state) => state.user.user.member_id);
 
 	const [page, setPage] = useState(1);
 
@@ -89,16 +87,8 @@ const CommentViewerContainer = () => {
 	};
 
 	return (
-		<CommentsViewer
-			articleId={articleId}
-			comments={comments}
-			recomments={recomments}
-			onCommentSubmit={onCommentSubmit}
-			onRecommentShow={onRecommentShow}
-		>
+		<CommentsViewer comments={comments}>
 			<CommentsList
-				comments={comments}
-				recomments={recomments}
 				onCommentSubmit={onCommentSubmit}
 				onRecommentShow={onRecommentShow}
 				onCommentDelete={onCommentDelete}
